Support RUST_BACKTRACE in test_fmt
Setting `RUST_BACKTRACE` when running tests is useful so that the logs
will contain backtraces when a test fails. `anyhow` uses this
environment variable as well when formatting its errors. `test_fmt`
currently tests that formatting is an exact match, so adding a trailing
backtrace causes them to fail.

This patch checks that the formatted error *starts with* the intended
result, thus allowing it to pass when a backtrace is appended.

Test: atest --host anyhow_host_test_tests_test_fmt w/RUST_BACKTRACE=full
Bug: 193841423
Change-Id: If52791f53e0deefea815c24cc4f7695201de7d79